pyproject.toml
[project]
name = "letsraid"
version = "0.1.0"
description = "Discord LFG bot for the First Wave Survivors server"
requires-python = ">=3.11"
dependencies = [
    "discord.py>=2.4",
    "aiosqlite>=0.20",
    "python-dotenv>=1.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
